requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
wechatpy
pycryptodome
anthropic
openai
requests
apscheduler
pydantic